rename pages -> options (#36)
